standardize Position, PositionId, and Offset to use usize/isize
- Change Position row/col from u8 to usize
- Change PositionId index from u8 to usize
- Change Offset deltas from i8 to isize
- Add PositionId::COUNT and PositionId::WIDTH constants
- Update PositionMap to use PositionId::COUNT
- Add Excel-style column labels (A-Z, AA-ZZ, AAA-ZZZ)

+++ src/cluster/strategy/mod.rs
@@ -97,7 +97,7 @@ mod tests {
     use crate::position::Position;
     use std::collections::HashSet;
 
-    fn pos(row: u8, col: u8) -> PositionId {
+    fn pos(row: usize, col: usize) -> PositionId {
         PositionId::from_position(Position::new(row, col))
     }
