+++ src/interactive_strategy.rs
@@ -102,7 +102,7 @@ mod tests {
     use crate::position::Position;
     use ratatui::backend::TestBackend;
 
-    fn pos(row: u8, col: u8) -> PositionId {
+    fn pos(row: usize, col: usize) -> PositionId {
         PositionId::from_position(Position::new(row, col))
     }
