+++ src/neighbor_cache.rs
@@ -57,7 +57,7 @@ mod tests {
     use super::*;
     use crate::position::Position;
 
-    fn pos(row: u8, col: u8) -> PositionId {
+    fn pos(row: usize, col: usize) -> PositionId {
         PositionId::from_position(Position::new(row, col))
     }
